Make LVM the default, and mount VGs as early as possible.
Unfortunately, we can't mount before the USBWAIT which is the proper way to wait long enough to allow our volume to appear. But we also don't want to wait the full 10 seconds for it to appear if we don't expect it to, so the I dropped the startup time to 2s.

One of the first things we now do in the image is activate the VGs so we can continue to use our attached drive for persistence of the Docker images and SSH keys.